LayoutTests/imported/w3c/web-platform-tests/web-share/share-unicode-strings-nonutf8-manual.https-expected.txt

Instructions:

Click the Share button.
Choose a valid share target.
Pass the test iff the target app received:

title = "fáncy 写作 😱"
text = "�x"
url = "https://xn--0zwm56d.example.com/%F0%9F%93%84"


Harness Error (TIMEOUT), message = null
